Placed handler functions in separate file
